fixed a few bugs more error toasts in process of moving the search to tab
