+++ app/src/main/res/layout/activity_search_movie.xml
@@ -11,32 +11,32 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="glassa.tacoma.uw.edu.moviesproject.search.SearchMovieActivity">
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:orientation="horizontal"
-        android:layout_height="wrap_content">
+    <!--<LinearLayout-->
+        <!--android:layout_width="match_parent"-->
+        <!--android:orientation="horizontal"-->
+        <!--android:layout_height="wrap_content">-->
 
 
-        <EditText
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:id="@+id/search_movie_bar"
-            android:hint="Find a movie..."
-            />
-        <Button
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight=".5"
-            android:id="@+id/search_movie_button"
-            android:text="Search" />
-    </LinearLayout>
+        <!--<EditText-->
+            <!--android:layout_width="0dp"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:layout_weight="1"-->
+            <!--android:id="@+id/search_movie_bar"-->
+            <!--android:hint="Find a movie..."-->
+            <!--/>-->
+        <!--<Button-->
+            <!--android:layout_width="0dp"-->
+            <!--android:layout_height="match_parent"-->
+            <!--android:layout_weight=".5"-->
+            <!--android:id="@+id/search_movie_button"-->
+            <!--android:text="Search" />-->
+    <!--</LinearLayout>-->
 
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:id="@+id/listview_search_user"/>
+    <!--<ListView-->
+        <!--android:layout_width="match_parent"-->
+        <!--android:layout_height="match_parent"-->
+        <!--android:layout_alignParentLeft="true"-->
+        <!--android:layout_alignParentStart="true"-->
+        <!--android:id="@+id/listview_search_user"/>-->
 
 </LinearLayout>
